vr_subscriber/requirements.txt
ratcave==1.0.0
scipy==1.5.2
PsychoPy==2020.2.10
PyQt5==5.15.2
paho-mqtt==1.5.1
fastlogging==0.9.5